requirements.txt
langchain==0.3.27
openai==2.1.0
langchain-openai
python-dotenv
langchain-community